add getKeys to map, move map to utils
